Lays out the ground work for responsiveness
I have set everything to have an `em` value. The pages media queries
simply changes the body font-size and the whole site updates spacing and
sizes accordingly, I like this, but it's tedious and possible will be
hard to change if we set the font sizes. I have not been able to achieve
the same thing using `rem`.